Puts ERIs in separate file
